+++ demos/04-vectorfield/index.js
@@ -49,54 +49,32 @@
 	    );
 	    pb.config.preDraw = function() { preDraw(); };
 
-	    // A list of vectors (position plus velocity)
 	    var pointlist = [];
-
-	    // +---------------------------------------------------------------------------------
-	    // | Clear and rebuild the point cloud with random points.
-	    // +-------------------------------
-	    var resetPointCloud = function() {
-			pointlist = [];
-			adjustPointCloud();
-	    };
-
-	    // +---------------------------------------------------------------------------------
-	    // | Expand or reduce the point cloud so the desired point count is obtained.
-	    // +-------------------------------
-	    var adjustPointCloud = function() {
-			var v;
-			while( pointlist.length < config.pointCount ) {
-				v = Vertex.randomVertex(pb.viewport());
-				pointlist.push( new Vector(v, v.clone()) );
-			}
-			if( pointlist.length > config.pointCount ) {
-				pointlist = pointlist.slice(0,config.pointCount);
-			}
-	    };
+	    for( var i = 0; i < 100; i++ ) {
+		pointlist.push( Vertex.randomVertex(pb.viewport()) );
+	    }
 	    
 
 	    // +---------------------------------------------------------------------------------
 	    // | Add a mouse listener to track the mouse position.
 	    // +-------------------------------
 	    new MouseHandler(pb.canvas)
-			.move( function(e) {
-				var relPos = pb.transformMousePosition( e.params.pos.x, e.params.pos.y );
-				var cx = document.getElementById('cx');
-				var cy = document.getElementById('cy');
-				if( cx ) cx.innerHTML = relPos.x.toFixed(2);
-				if( cy ) cy.innerHTML = relPos.y.toFixed(2);
-			} );
+		.move( function(e) {
+		    var relPos = pb.transformMousePosition( e.params.pos.x, e.params.pos.y );
+		    var cx = document.getElementById('cx');
+		    var cy = document.getElementById('cy');
+		    if( cx ) cx.innerHTML = relPos.x.toFixed(2);
+		    if( cy ) cy.innerHTML = relPos.y.toFixed(2);
+		} );
 
 
 	    // +---------------------------------------------------------------------------------
 	    // | A global config that's attached to the dat.gui control interface.
 	    // +-------------------------------
 	    var config = PlotBoilerplate.utils.safeMergeByKeys( {
-			showFieldVectors      : true,
-			animate               : false,
-			pointCount            : 100,
-			resetPointCloud       : function() { resetPointCloud(); }
-			}, GUP );
+		showFieldVectors      : true,
+		animate               : false
+	    }, GUP );
 	    
 
 	    // +---------------------------------------------------------------------------------
@@ -111,6 +89,7 @@
 	    pb.add( cv3 = new Vector( new Vertex(controlOffset,-controlOffset), new Vertex(controlOffset+diameter*0.6,-controlOffset-diameter*0.6) ), false );
 	    pb.add( cv4 = new Vector( new Vertex(controlOffset,controlOffset), new Vertex(controlOffset+diameter*0.6,controlOffset+diameter*0.6) ), false );
 	    var controlVectors = [ cv0, cv1, cv2, cv3, cv4 ];
+	    //var field = [];
 	    
 	    
 	    // +--------------------------------------------------------------------------------
@@ -121,82 +100,129 @@
 	    var displayStep = 25;
 
 	    function drawFieldVectorAt( vertex ) {
-			var vector = new Vector( vertex, vertex.clone() );
-			for( var ci in controlVectors ) {
-				var controlVector = controlVectors[ci];
-				var strength = controlVector.length();
-				
-				adjustVector_radial( controlVector, strength, vector );
-			}
-			if( vector.length() > 0 )
-				pb.draw.arrow( vector.a, vector.b, '#e888e8' );
-			else
-				pb.draw.dot( vector.a, '#e800e8' );
+		var vector = new Vector( vertex, vertex.clone() );
+		for( var ci in controlVectors ) {
+		    var controlVector = controlVectors[ci];
+		    var strength = controlVector.length();
+		    
+		    adjustVector_radial( controlVector, strength, vector );
+		}
+		if( vector.length() > 0 )
+		    pb.draw.arrow( vector.a, vector.b, '#e888e8' );
+		else
+		    pb.draw.dot( vector.a, '#e800e8' );
 	    }
 	    function preDraw() {
-			if( config.showFieldVectors ) {
-				for( var x = 0; x < pb.canvasSize.width/2; x+=displayStep ) {
-				for( var y = 0; y < pb.canvasSize.height/2; y+=displayStep ) {
-					drawFieldVectorAt( vert(x,y) );
-					drawFieldVectorAt( vert(-x,y) );
-					drawFieldVectorAt( vert(x,-y) );
-					drawFieldVectorAt( vert(-x,-y) );
-				}
-				}
+		//console.log( 'preDraw' );		    
+		if( config.showFieldVectors ) {
+		    //console.log( 'draw field vectors' );
+		    for( var x = 0; x < pb.canvasSize.width/2; x+=displayStep ) {
+			for( var y = 0; y < pb.canvasSize.height/2; y+=displayStep ) {
+			    drawFieldVectorAt( vert(x,y) );
+			    drawFieldVectorAt( vert(-x,y) );
+			    drawFieldVectorAt( vert(x,-y) );
+			    drawFieldVectorAt( vert(-x,-y) );
 			}
-			if( config.animate ) {
-				for( var i in pointlist ) {
-				pb.draw.arrow( pointlist[i].a, pointlist[i].b, 'black' );
-				pb.fill.circle( pointlist[i].a, 2, 'black' );
-				}
-			}
+		    }
+		}
+		if( config.animate ) {
+		    //console.log('draw point cloud');
+		    for( var i in pointlist ) {
+			pb.fill.circle( pointlist[i], 2, 'black' );
+		    }
+		}
 		    
 	    };
 
-	    function adjustPointList( stepFactor ) {
-			for( var fi in pointlist ) {
-				var pointWithVelocity = pointlist[fi];
-				// var anyVector = pointWithVelocity.clone();
-				anyVector.b.set( anyVector.a );
-				
-				for( var ci in controlVectors ) {
-					var controlVector = controlVectors[ci];
-					var strength = controlVector.length();
-					var anyVector = pointWithVelocity.clone();
-					adjustVector_radial( controlVector, strength, anyVector );
+	    /*
+	    function adjustVectorField() {
+		 for( var fi in vectorField ) {
+		     var anyVector = vectorField[fi];
+		     // Set to length 0
+		     anyVector.b.set(anyVector.a);
+		     
+		     for( var ci in controlVectors ) {
+			 var controlVector = controlVectors[ci];
+			 var strength = controlVector.length();
+			 adjustVector_radial( controlVector, strength, anyVector );
 
-					// Relocate vector to a=(0,0), but keep an only copy.
-					pointWithVelocity.a.add( anyVector.clone().sub(anyVector.a).b.multiplyScalar(stepFactor) );
-					pointWithVelocity.b.set( anyVector.b );
-				}
-			}
-			pb.redraw();
+		     }
 		}
+		pb.redraw();
+	    }
+	    */
 
-		function adjustVector_radial( controlVector, strength, anyVector ) {
-			var dist = controlVector.a.distance( anyVector.a );
-			if( dist > strength )
-				return;
-			var influence = dist-strength;
-			anyVector.b.x += (controlVector.a.x-controlVector.b.x)*(influence/strength);
-			anyVector.b.y += (controlVector.a.y-controlVector.b.y)*(influence/strength);
-			
+	     function adjustPointList( stepFactor ) {
+		 for( var fi in pointlist ) {
+		     var point = pointlist[fi];
+		     var anyVector = new Vector( point.clone(), point.clone() );
+		     
+		     for( var ci in controlVectors ) {
+			 var controlVector = controlVectors[ci];
+			 var strength = controlVector.length();
+			 adjustVector_radial( controlVector, strength, anyVector );
+
+		     }
+
+		     // Relocate vector to a=(0,0)
+		     anyVector.sub( anyVector.a );
+		     pointlist[fi].add( anyVector.b.multiplyScalar(stepFactor) );
 		}
+		pb.redraw();
+	    }
 
-		function renderAnimation() {
-			//adjustVectorField();
-			adjustPointList( 0.01 );
-			//pb.redraw();
-			if( config.animate )
-				window.requestAnimationFrame( renderAnimation );
-			};
-			
-			function toggleAnimation() {
-			if( config.animate ) {
-				renderAnimation();
-			} else {
-				pb.redraw();
-			}
+	    /*
+	      // Some test which still looks nice
+	    function adjustVector_radial( controlVector, strength, anyVector ) {
+		var t = controlVector.getClosestT(anyVector.a);
+		// t = Math.max(0, Math.min(1,t));
+		var pointOnControlVector = controlVector.vertAt(t);
+		var dist = pointOnControlVector.distance( anyVector.a );
+		// Is the vertex inside the strengh area of thi control Vector?
+		if( dist < strength ) {
+		    // console.log('x');
+
+		    // dist -= strength;
+		    anyVector.b.x += (anyVector.a.x-controlVector.a.x)*(strength/dist);
+		    anyVector.b.y += (anyVector.a.y-controlVector.a.y)*(strength/dist);
+		    //anyVector.b.x += (anyVector.a.x-controlVector.a.x)*(dist/strength);
+		    //anyVector.b.y += (anyVector.a.y-controlVector.a.y)*(dist/strength);
+		}
+	    }
+	    */
+
+	    function adjustVector_radial( controlVector, strength, anyVector ) {
+		var dist = controlVector.a.distance( anyVector.a );
+		if( dist > strength )
+		    return;
+		var influence = dist-strength;
+		anyVector.b.x += (controlVector.a.x-controlVector.b.x)*(influence/strength);
+		anyVector.b.y += (controlVector.a.y-controlVector.b.y)*(influence/strength);
+		
+	    }
+
+	    // Install listeners to all control vectors
+	    /*for( var i in controlVectors ) {
+		controlVectors[i].a.listeners.addDragListener( function(e) { adjustVectorField(); } );
+		controlVectors[i].b.listeners.addDragListener( function(e) { adjustVectorField(); } );
+	    }
+	    adjustVectorField();
+	    */
+
+
+	    function renderAnimation() {
+		//adjustVectorField();
+		adjustPointList( 0.01 );
+		//pb.redraw();
+		window.requestAnimationFrame( renderAnimation );
+	    };
+	    
+	    function toggleAnimation() {
+		if( config.animate ) {
+		    renderAnimation();
+		} else {
+		    pb.redraw();
+		}
 	    }
 
 	    // +---------------------------------------------------------------------------------
@@ -207,12 +233,9 @@
 		var f0 = gui.addFolder('Points');
 		f0.add(config, 'showFieldVectors').onChange( function() { pb.redraw(); } ).name('Show field vectors').title('Show field vectors.');
 		f0.add(config, 'animate').onChange( function() { toggleAnimation(); } ).name('Animate a point cloud').title('Animate a point cloud.');
-		f0.add(config, 'pointCount').min(1).max(1000).onChange( function() { adjustPointCloud(); } ).name('The number of points you want to animate').title('The number of points you want to animate.');
-		f0.add(config, 'resetPointCloud').name('Reset point cloud').title('Reset point cloud.');
 		f0.open();
 	    }
 
-	    resetPointCloud();
 	    toggleAnimation();
 
 	} );